- now blocking wrong cf handles from being added - moved cfAPI functions to a separate file 	Resolves #11
